# docs/topics/gamification_openapi_2.self-contained.yaml
...
openapi: "3.0.3"
info:
  title: "gamification API"
  description: "gamification API"
  version: "1.0.0"
servers:
  - url: "https://api.tc-syd-02.insa-lyon.fr"
    description: "Production server"
paths:
...
  /api/login:
    post:
      summary: "Login"
      description: ""
      security: [ ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "username"
                password:
                  type: string
                  example: "password"
      responses:
        "200":
          description: >
            Successfully authenticated.
            The session ID is returned in a cookie named `connect.sid`. You need to include this cookie in subsequent requests.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: connect.sid=s%3AVwb4Nr25tPu4LgLMkaH123_Zsm6BcE0t.cHjERY5R7l2obZeQgDuBYTOj4MZwEa59VM0k3cx%2Frog; Path=/; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Successfully authenticated"
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid username or password"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/logout:
    post:
      summary: "Logout"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "Ok"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/register:
    post:
      summary: "Register a new user"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "username"
                password:
                  type: string
                  example: "password"
                name:
                  type: string
                  example: "name"
                surname:
                  type: string
                  example: "surname"
                email:
                  type: string
                  example: "email"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Missing required fields"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
...
components:
  schemas:
...
    User:
      description: "A user"
      type: "object"
      properties:
        id:
          description: "The user's id"
          type: "string"
        username:
          description: "The user's username"
          type: "string"
        name:
          description: "The user's name"
          type: "string"
        surname:
          description: "The user's surname"
          type: "string"
        email:
          description: "The user's email"
          type: "string"
        type:
          description: "The user's type"
          type: "string"
          enum:
            - "student"
            - "teacher"
            - "admin"
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid